+++ crates/forge/src/run_tests.rs
@@ -1,6 +1,6 @@
 pub mod maat;
+pub mod messages;
 pub mod package;
 pub mod resolve_config;
-pub mod structs;
 pub mod test_target;
 pub mod workspace;
