+++ crates/forge/src/run_tests/messages/mod.rs
@@ -1,0 +1,6 @@
+pub mod collected_tests_count;
+pub mod latest_blocks_numbers;
+pub mod overall_summary;
+pub mod tests_failure_summary;
+pub mod tests_run;
+pub mod tests_summary;
