+++ crates/forge/src/run_tests/test_target.rs
@@ -1,5 +1,6 @@
 use anyhow::Result;
 use forge_runner::messages::TestResultMessage;
+
 use forge_runner::{
     TestCaseFilter,
     forge_config::ForgeConfig,
